Clean up JavaScript files.
