+++ app/javascript/spree/dashboard/index.js
@@ -1,13 +1,32 @@
-import "./controllers"
-import "./utilities/bootstrap"
-
+//
+// Import JavaScript packages that are required globally.
+import { Application } from "@hotwired/stimulus"
 import flatpickr from "flatpickr"
 import * as Turbo from "@hotwired/turbo"
 
-class Dashboard {
-  constructor() {
-    console.log("Spree Dashboard Initiated")
-  }
-}
 
-export { Dashboard, flatpickr, Turbo }
+//
+// Import Utility JavaScript required globally.
+import "./utilities/bootstrap"
+
+
+//
+// Stimulus - Setup
+const application = Application.start()
+application.debug = false
+window.Stimulus = application
+
+// Stimulus - Controllers
+import UploadButtonController from "./controllers/upload_button_controller"
+application.register("upload-button", UploadButtonController)
+
+import SpreeController from "./controllers/spree_controller"
+application.register("spree", SpreeController)
+
+import SortableTreeController from "./controllers/sortable_tree_controller"
+application.register("sortable-tree", SortableTreeController)
+
+
+//
+// Export
+export { application, flatpickr, Turbo }
